+++ cuda_src/ALS_onCUDA.h
@@ -17,11 +17,11 @@ __device__ void choldcsl_k(int n, float** A);
 
 __device__ void inverseMatrix_CholeskyMethod_k(int n, float** A);
 
-__device__ void Mt_byM_multiply_k(int i, int j, float* H, float** Result, long ptr, const unsigned* idx);
+__device__ void Mt_byM_multiply_k(long i, long j, float* H, float** Result, long ptr, const long* idx);
 
-__global__ void updateW_overH_kernel(long rows, const long* row_ptr, const unsigned* col_idx, const unsigned* colMajored_sparse_idx, const float* val, float lambda, unsigned k, float* W, float* H);
+__global__ void updateW_overH_kernel(long rows, const long* row_ptr, const long* col_idx, const long* colMajored_sparse_idx, const float* val, float lambda, unsigned k, float* W, float* H);
 
-__global__ void updateH_overW_kernel(long cols, const long* col_ptr, const unsigned* row_idx, const float* val, float lambda, unsigned k, float* W, float* H);
+__global__ void updateH_overW_kernel(long cols, const long* col_ptr, const long* row_idx, const float* val, float lambda, unsigned k, float* W, float* H);
 
 void kernel_wrapper_als_NV(smat_t& R, testset_t& T, mat_t& W, mat_t& H, parameter& parameters);
